parse tests into multiple files and add stages and jobs to travis (#32)
* parse tests into multiple files

* add stages and jobs to travis

* modify travis.yml - could not parse

* remove double PR testing
