+++ .travis.yml
@@ -1,4 +1,16 @@
 language: python
 
-# command to run tests
-script: pytest
+stages:
+  - name: basic tests
+    if: type = push
+  - name: behind Earthdata
+    if: branch = master OR commit_message =~ nsidc_tests OR type = cron
+
+#ultimately move basic tests list into a separate script with the list of files to run (then this can be put into a matrix and paralellized as well)
+jobs:
+  include:
+    - stage: basic tests
+      script: pytest icepyx/tests/test_is2class_inputs.py icepyx/tests/test_is2class_query.py icepyx/tests/test_is2class_methods.py
+        
+    - stage: behind Earthdata
+      script: pytest icepyx/tests/test_is2class_NSIDC-API.py
